Modify implementation of merge
In last version of merge, the implementation is to insert the node from
right list into left list one by one, which is asymmetrical, needs to
consider special case and hard to read. Therefore, I modify the
implementation in this version. First deciding the head of list, and
determine the next node inserting into head from left list or right
list. I think this implementation is more readable and concise.