Add parameter compare_count to qtest
To count the compare times in sorting, I add a global variable
cmp_count in queue.c. When q_sort or list_sort being used, it will set
the cmp_count to zero, and increment it each time compare occurs.
When the sorting is done, use command "option" to show the parameters,
and cmp_count will be printed on the screen.
The reason I add this parameter is to analysis the performance of
sorting algorithm.