Add command average_k
Design a experiment to measure the comparison times. K is a function
of fraction part of log2(n). We have a mathematical formula

 - cmp_count = n*log2(n) - K*n + 1

so

 - K = log2(n) - (cmp_count-1)/n

The command average_k have to parameter, first is sort / kernel_sort,
which is use to determine which sorting we want to use. Second is a
number n, n is a power of two.The reason for power of two is that K is
 a function of fraction part, which means it a periodic function of
log2(n), so we only need to compute average by range from n to 2*n-1
where n power of two.